cleanup: split files + gobindings
